feat(MAR-183): Implement NotificationManager, double-opt-in reminders, and last-minute booking restrictions
